refactor: group test utils
