+++ src/test/java/gdsc/konkuk/platformcore/application/email/EmailTaskFacadeTest.java
@@ -12,8 +12,8 @@ import gdsc.konkuk.platformcore.domain.email.entity.EmailDetail;
 import gdsc.konkuk.platformcore.domain.email.entity.EmailReceiver;
 import gdsc.konkuk.platformcore.domain.email.entity.EmailReceivers;
 import gdsc.konkuk.platformcore.domain.email.entity.EmailTask;
-import gdsc.konkuk.platformcore.fixture.email.EmailSendRequestFixture;
-import gdsc.konkuk.platformcore.fixture.email.EmailTaskFixture;
+import gdsc.konkuk.platformcore.util.fixture.email.EmailSendRequestFixture;
+import gdsc.konkuk.platformcore.util.fixture.email.EmailTaskFixture;
 import gdsc.konkuk.platformcore.global.scheduler.TaskScheduler;
 import java.time.LocalDateTime;
 import java.util.Set;
